Refactor client tests so mocks can be shared
